♻️ move sendWithRetryStrategy and Payload types to @datadog/js-core/transport
