+++ packages/browser-core/src/transport/batch.spec.ts
@@ -1,3 +1,4 @@
+import type { HttpRequestEvent } from '@datadog/js-core/transport'
 import { Observable } from '..'
 import type { MockFlushController } from '../../test'
 import { createMockFlushController, replaceMockable } from '../../test'
@@ -8,7 +9,7 @@ import { createPageMayExitObservable } from '../browser/pageMayExitObservable'
 import { createBatch, MESSAGE_BYTES_LIMIT, type Batch } from './batch'
 import { createFlushController } from './flushController'
 import { createHttpRequest } from './httpRequest'
-import type { HttpRequest, HttpRequestEvent } from './httpRequest'
+import type { HttpRequest } from './httpRequest'
 
 describe('batch', () => {
   const BIG_MESSAGE_OVER_BYTES_LIMIT = { message: 'x'.repeat(MESSAGE_BYTES_LIMIT + 1) }
