+++ packages/browser-core/src/transport/httpRequest.spec.ts
@@ -1,4 +1,4 @@
-import type { EndpointBuilder } from '@datadog/js-core/transport'
+import type { EndpointBuilder, HttpRequestEvent  } from '@datadog/js-core/transport'
 import { createEndpointBuilder } from '@datadog/js-core/transport'
 import type { Request } from '../../test'
 import {
@@ -11,7 +11,7 @@ import {
   wait,
 } from '../../test'
 import { noop } from '../tools/utils/functionUtils'
-import type { HttpRequest, HttpRequestEvent } from './httpRequest'
+import type { HttpRequest } from './httpRequest'
 import { createHttpRequest, fetchStrategy, RECOMMENDED_REQUEST_BYTES_LIMIT } from './httpRequest'
 
 describe('httpRequest', () => {
