+++ packages/js-core/src/entries/transport.ts
@@ -11,6 +11,17 @@ export type {
 export { createEndpointBuilder, createReplicaEndpointBuilder, buildEndpointUrl } from '../transport/endpointBuilder'
 export type { Encoder, EncoderResult } from '../transport/encoder'
 export { createIdentityEncoder } from '../transport/encoder'
+export type { Payload, HttpResponse, BandwidthStats, HttpRequestEvent } from '../transport/payload'
+export {
+  sendWithRetryStrategy,
+  newRetryState,
+  MAX_ONGOING_BYTES_COUNT,
+  MAX_ONGOING_REQUESTS,
+  MAX_QUEUE_BYTES_COUNT,
+  MAX_BACKOFF_TIME,
+  INITIAL_BACKOFF_TIME,
+} from '../transport/sendWithRetryStrategy'
+export type { RetryState } from '../transport/sendWithRetryStrategy'
 export type { Site } from '../transport/intakeSites'
 export {
   INTAKE_SITE_STAGING,
